moleculesに分割
