+++ frontend/app/nextjs-blog/components/molecules/author/index.js
@@ -1,0 +1,81 @@
+export default function Author() {
+	return (
+		<ul className="-mx-4">
+			<li className="flex items-center">
+				<img
+					src="https://images.unsplash.com/photo-1492562080023-ab3db95bfbce?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=731&q=80"
+					alt="avatar"
+					className="object-cover w-10 h-10 mx-4 rounded-full"
+				/>
+				<p>
+					<a href="#" className="mx-1 font-bold text-gray-700 hover:underline">
+						Alex John
+					</a>
+					<span className="text-sm font-light text-gray-700">
+						Created 23 Posts
+					</span>
+				</p>
+			</li>
+			<li className="flex items-center mt-6">
+				<img
+					src="https://images.unsplash.com/photo-1464863979621-258859e62245?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=333&q=80"
+					alt="avatar"
+					className="object-cover w-10 h-10 mx-4 rounded-full"
+				/>
+				<p>
+					<a href="#" className="mx-1 font-bold text-gray-700 hover:underline">
+						Jane Doe
+					</a>
+					<span className="text-sm font-light text-gray-700">
+						Created 52 Posts
+					</span>
+				</p>
+			</li>
+			<li className="flex items-center mt-6">
+				<img
+					src="https://images.unsplash.com/photo-1531251445707-1f000e1e87d0?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=281&q=80"
+					alt="avatar"
+					className="object-cover w-10 h-10 mx-4 rounded-full"
+				/>
+				<p>
+					<a href="#" className="mx-1 font-bold text-gray-700 hover:underline">
+						Lisa Way
+					</a>
+					<span className="text-sm font-light text-gray-700">
+						Created 73 Posts
+					</span>
+				</p>
+			</li>
+			<li className="flex items-center mt-6">
+				<img
+					src="https://images.unsplash.com/photo-1500757810556-5d600d9b737d?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=735&q=80"
+					alt="avatar"
+					className="object-cover w-10 h-10 mx-4 rounded-full"
+				/>
+				<p>
+					<a href="#" className="mx-1 font-bold text-gray-700 hover:underline">
+						Steve Matt
+					</a>
+					<span className="text-sm font-light text-gray-700">
+						Created 245 Posts
+					</span>
+				</p>
+			</li>
+			<li className="flex items-center mt-6">
+				<img
+					src="https://images.unsplash.com/photo-1502980426475-b83966705988?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=373&q=80"
+					alt="avatar"
+					className="object-cover w-10 h-10 mx-4 rounded-full"
+				/>
+				<p>
+					<a href="#" className="mx-1 font-bold text-gray-700 hover:underline">
+						Khatab Wedaa
+					</a>
+					<span className="text-sm font-light text-gray-700">
+						Created 332 Posts
+					</span>
+				</p>
+			</li>
+		</ul>
+	);
+}
