+++ frontend/app/nextjs-blog/components/molecules/category/index.js
@@ -1,0 +1,54 @@
+export default function Category() {
+	return (
+		<ul>
+			<li>
+				<a
+					href="#"
+					className="mx-1 font-bold text-gray-700 hover:text-gray-600 hover:underline"
+				>
+					- AWS
+				</a>
+			</li>
+			<li className="mt-2">
+				<a
+					href="#"
+					className="mx-1 font-bold text-gray-700 hover:text-gray-600 hover:underline"
+				>
+					- Laravel
+				</a>
+			</li>
+			<li className="mt-2">
+				<a
+					href="#"
+					className="mx-1 font-bold text-gray-700 hover:text-gray-600 hover:underline"
+				>
+					- Vue
+				</a>
+			</li>
+			<li className="mt-2">
+				<a
+					href="#"
+					className="mx-1 font-bold text-gray-700 hover:text-gray-600 hover:underline"
+				>
+					- Design
+				</a>
+			</li>
+			<li className="flex items-center mt-2">
+				<a
+					href="#"
+					className="mx-1 font-bold text-gray-700 hover:text-gray-600 hover:underline"
+				>
+					- Django
+				</a>
+			</li>
+			<li className="flex items-center mt-2">
+				<a
+					href="#"
+					className="mx-1 font-bold text-gray-700 hover:text-gray-600 hover:underline"
+				>
+					- PHP
+				</a>
+			</li>
+		</ul>
+	);
+}
