+++ frontend/app/nextjs-blog/components/molecules/post/index.js
@@ -1,16 +1,7 @@
-export default function Posts() {
-  return (
-		<div className="w-full lg:w-8/12">
-			<div className="flex items-center justify-between">
-				<h1 className="text-xl font-bold text-gray-700 md:text-2xl">Post</h1>
-				<div>
-					<select className="w-full border-gray-300 rounded-md shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50">
-						<option>Latest</option>
-						<option>Last Week</option>
-					</select>
-				</div>
-			</div>
-			<div className="mt-6">
+export default function Post(){
+return(
+  <div>
+  <div className="mt-6">
 				<div className="max-w-4xl px-10 py-6 mx-auto bg-white rounded-lg shadow-md">
 					<div className="flex items-center justify-between">
 						<span className="font-light text-gray-600">Jun 1, 2020</span>
@@ -229,41 +220,8 @@ export default function Posts() {
 						</div>
 					</div>
 				</div>
-			</div>
-			<div className="mt-8">
-				<div className="flex">
-					<a
-						href="#"
-						className="px-3 py-2 mx-1 font-medium text-gray-500 bg-white rounded-md cursor-not-allowed"
-					>
-						previous
-					</a>
-					<a
-						href="#"
-						className="px-3 py-2 mx-1 font-medium text-gray-700 bg-white rounded-md hover:bg-blue-500 hover:text-white"
-					>
-						1
-					</a>
-					<a
-						href="#"
-						className="px-3 py-2 mx-1 font-medium text-gray-700 bg-white rounded-md hover:bg-blue-500 hover:text-white"
-					>
-						2
-					</a>
-					<a
-						href="#"
-						className="px-3 py-2 mx-1 font-medium text-gray-700 bg-white rounded-md hover:bg-blue-500 hover:text-white"
-					>
-						3
-					</a>
-					<a
-						href="#"
-						className="px-3 py-2 mx-1 font-medium text-gray-700 bg-white rounded-md hover:bg-blue-500 hover:text-white"
-					>
-						Next
-					</a>
-				</div>
-			</div>
-		</div>
-	);
-}
+      </div>
+    </div>
+)
+
+};
