+++ frontend/app/nextjs-blog/components/organisms/footer.js
